move pytest tests to different files to run in separate processes
and move a few functions to tests/utils.py
and update the .travis.yml to run the tests as separate processes
this should solve a bug where training sessions aren't cleaned properly
and concurrent trainings result in different results
